Web.Scotty.Internal.Types

Documentation

data Options

Constructors

Options

Fields

verbose :: Int

0 = silent, 1(def) = startup banner

settings :: Settings

Warp Settings Note: to work around an issue in warp, the default FD cache duration is set to 0 so changes to static files are always picked up. This likely has performance implications, so you may want to modify this for production servers using setFdCacheDuration.

defaultOptions :: Options

newtype RouteOptions

Constructors

RouteOptions

Fields

maxRequestBodySize :: Maybe Kilobytes

defaultRouteOptions :: RouteOptions

type Kilobytes = Int

type Middleware m = Application m -> Application m

type Application m = Request -> m Response

data BodyChunkBuffer

Constructors

BodyChunkBuffer

Fields

hasFinishedReadingChunks :: Bool

whether we've reached the end of the stream yet

chunksReadSoFar :: [ByteString]

data BodyInfo

The key part of having two MVars is that we can "clone" the BodyInfo to create a copy where the index is reset to 0, but the chunk cache is the same. Passing a cloned BodyInfo into each matched route allows them each to start from the first chunk if they call bodyReader.

Introduced in (#308)

Constructors

BodyInfo

Fields

bodyInfoReadProgress :: MVar Int

index into the stream read so far

bodyInfoChunkBuffer :: MVar BodyChunkBuffer

bodyInfoDirectChunkRead :: IO ByteString

can be called to get more chunks

data ScottyState m

Constructors

ScottyState

Fields

middlewares :: [Middleware]

routes :: [BodyInfo -> Middleware m]

handler :: Maybe (ErrorHandler m)

routeOptions :: RouteOptions

defaultScottyState :: ScottyState m

addMiddleware :: Middleware -> ScottyState m -> ScottyState m

addRoute :: (BodyInfo -> Middleware m) -> ScottyState m -> ScottyState m

setHandler :: Maybe (ErrorHandler m) -> ScottyState m -> ScottyState m

updateMaxRequestBodySize :: RouteOptions -> ScottyState m -> ScottyState m

newtype ScottyT m a

Constructors

ScottyT

Fields

runS :: State (ScottyState m) a

data ActionError

Internal exception mechanism used to modify the request processing flow.

The exception constructor is not exposed to the user and all exceptions of this type are caught and processed within the runAction function.

Constructors

AERedirect Text
Redirect

AENext
Stop processing this route and skip to the next one

AEFinish
Stop processing the request

tryNext :: MonadUnliftIO m => m a -> m Bool

data StatusError

Deprecated: If it is supposed to be caught, a proper exception type should be defined

E.g. when a parameter is not found in a query string (400 Bad Request) or when parsing a JSON body fails (422 Unprocessable Entity)

Constructors

StatusError Status Text

Deprecated: If it is supposed to be caught, a proper exception type should be defined

type ErrorHandler m = Handler (ActionT m) ()

Specializes a Handler to the ActionT monad

data ScottyException

Thrown e.g. when a request is too large

Constructors

RequestTooLarge

MalformedJSON ByteString Text

FailedToParseJSON ByteString Text

PathParameterNotFound Text

QueryParameterNotFound Text

FormFieldNotFound Text

FailedToParseParameter Text Text Text

type Param = (Text, Text)

type File = (Text, FileInfo ByteString)

data ActionEnv

Constructors

Env

Fields

envReq :: Request

envPathParams :: [Param]

envFormParams :: [Param]

envQueryParams :: [Param]

envBody :: IO ByteString

envBodyChunk :: IO ByteString

envFiles :: [File]

envResponse :: TVar ScottyResponse

getResponse :: MonadIO m => ActionEnv -> m ScottyResponse

getResponseAction :: MonadIO m => ActionT m ScottyResponse

modifyResponse :: MonadIO m => (ScottyResponse -> ScottyResponse) -> ActionT m ()

data BodyPartiallyStreamed

Constructors

BodyPartiallyStreamed

data Content

Constructors

ContentBuilder Builder

ContentFile FilePath

ContentStream StreamingBody

ContentResponse Response

data ScottyResponse

Constructors

SR

Fields

srStatus :: Status

srHeaders :: ResponseHeaders

srContent :: Content

setContent :: Content -> ScottyResponse -> ScottyResponse

setHeaderWith :: ([(HeaderName, ByteString)] -> [(HeaderName, ByteString)]) -> ScottyResponse -> ScottyResponse

setStatus :: Status -> ScottyResponse -> ScottyResponse

defaultScottyResponse :: ScottyResponse

The default response has code 200 OK and empty body

newtype ActionT m a

Constructors

ActionT

Fields

runAM :: ReaderT ActionEnv m a

tryAnyStatus :: MonadUnliftIO m => m a -> m Bool

catches either ActionError (thrown by next), ScottyException (thrown if e.g. a query parameter is not found) or StatusError (via raiseStatus)

data RoutePattern

Constructors

Capture Text

Literal Text

Function (Request -> Maybe [Param])