Web.Scotty.Trans

Description

It should be noted that most of the code snippets below depend on the OverloadedStrings language pragma.

The functions in this module allow an arbitrary monad to be embedded in Scotty's monad transformer stack, e.g. for complex endpoint configuration, interacting with databases etc.

Scotty is set up by default for development mode. For production servers, you will likely want to modify settings and the defaultHandler. See the comments on each of these functions for more information.

Please refer to the examples directory and the spec test suite for concrete use cases, e.g. constructing responses, exception handling and useful implementation details.

Running scotty servers

scottyT

Arguments

:: (Monad m, MonadIO n)

=> Port

-> (m Response -> IO Response)
Run monad m into IO, called at each action.

-> ScottyT m ()

-> n ()

Run a scotty application using the warp server. NB: scotty p === scottyT p id

scottyOptsT

Arguments

:: (Monad m, MonadIO n)

=> Options

-> (m Response -> IO Response)
Run monad m into IO, called at each action.

-> ScottyT m ()

-> n ()

Run a scotty application using the warp server, passing extra options. NB: scottyOpts opts === scottyOptsT opts id

scottySocketT :: (Monad m, MonadIO n) => Options -> Socket -> (m Response -> IO Response) -> ScottyT m () -> n ()

Run a scotty application using the warp server, passing extra options, and listening on the provided socket. NB: scottySocket opts sock === scottySocketT opts sock id

data Options

Constructors

Options

Fields

verbose :: Int

0 = silent, 1(def) = startup banner

settings :: Settings

Warp Settings Note: to work around an issue in warp, the default FD cache duration is set to 0 so changes to static files are always picked up. This likely has performance implications, so you may want to modify this for production servers using setFdCacheDuration.

defaultOptions :: Options

scotty-to-WAI

scottyAppT

Arguments

:: (Monad m, Monad n)

=> (m Response -> IO Response)
Run monad m into IO, called at each action.

-> ScottyT m ()

-> n Application

Turn a scotty application into a WAI Application, which can be run with any WAI handler. NB: scottyApp === scottyAppT id

Defining Middleware and Routes

Middleware and routes are run in the order in which they are defined. All middleware is run first, followed by the first route that matches. If no route matches, a 404 response is given.

middleware :: Middleware -> ScottyT m ()

Use given middleware. Middleware is nested such that the first declared is the outermost middleware (it has first dibs on the request and last action on the response). Every middleware is run on each request.

get :: MonadUnliftIO m => RoutePattern -> ActionT m () -> ScottyT m ()

get = addroute GET

post :: MonadUnliftIO m => RoutePattern -> ActionT m () -> ScottyT m ()

post = addroute POST

put :: MonadUnliftIO m => RoutePattern -> ActionT m () -> ScottyT m ()

put = addroute PUT

delete :: MonadUnliftIO m => RoutePattern -> ActionT m () -> ScottyT m ()

delete = addroute DELETE

patch :: MonadUnliftIO m => RoutePattern -> ActionT m () -> ScottyT m ()

patch = addroute PATCH

options :: MonadUnliftIO m => RoutePattern -> ActionT m () -> ScottyT m ()

options = addroute OPTIONS

addroute :: MonadUnliftIO m => StdMethod -> RoutePattern -> ActionT m () -> ScottyT m ()

Define a route with a StdMethod, a route pattern representing the path spec, and an Action which may modify the response.

get "/" $ text "beam me up!"

The path spec can include values starting with a colon, which are interpreted as captures. These are parameters that can be looked up with pathParam.

>>> :{
let server = S.get "/foo/:bar" (S.pathParam "bar" >>= S.text)
 in do
      withScotty server $ curl "http://localhost:3000/foo/something"
:}
"something"

matchAny :: MonadUnliftIO m => RoutePattern -> ActionT m () -> ScottyT m ()

Add a route that matches regardless of the HTTP verb.

notFound :: MonadUnliftIO m => ActionT m () -> ScottyT m ()

Specify an action to take if nothing else is found. Note: this _always_ matches, so should generally be the last route specified.

setMaxRequestBodySize

Arguments

:: Kilobytes
Request size limit

-> ScottyT m ()

Set global size limit for the request body. Requests with body size exceeding the limit will not be processed and an HTTP response 413 will be returned to the client. Size limit needs to be greater than 0, otherwise the application will terminate on start.

Route Patterns

capture :: String -> RoutePattern

Standard Sinatra-style route. Named captures are prepended with colons. This is the default route type generated by OverloadedString routes. i.e.

get (capture "/foo/:bar") $ ...

and

{-# LANGUAGE OverloadedStrings #-}
...
get "/foo/:bar" $ ...

are equivalent.

regex :: String -> RoutePattern

Match requests using a regular expression. Named captures are not yet supported.

>>> :{
let server = S.get (S.regex "^/f(.*)r$") $ do
                cap <- S.pathParam "1"
                S.text cap
 in do
      withScotty server $ curl "http://localhost:3000/foo/bar"
:}
"oo/ba"

function :: (Request -> Maybe [Param]) -> RoutePattern

Build a route based on a function which can match using the entire Request object. Nothing indicates the route does not match. A Just value indicates a successful match, optionally returning a list of key-value pairs accessible by param.

>>> :{
let server = S.get (function $ \req -> Just [("version", T.pack $ show $ W.httpVersion req)]) $ do
                v <- S.pathParam "version"
                S.text v
 in do
      withScotty server $ curl "http://localhost:3000/"
:}
"HTTP/1.1"

literal :: String -> RoutePattern

Build a route that requires the requested path match exactly, without captures.

Accessing the Request and its fields

request :: Monad m => ActionT m Request

Get the Request object.

header :: Monad m => Text -> ActionT m (Maybe Text)

Get a request header. Header name is case-insensitive.

headers :: Monad m => ActionT m [(Text, Text)]

Get all the request headers. Header names are case-insensitive.

body :: MonadIO m => ActionT m ByteString

Get the request body.

bodyReader :: Monad m => ActionT m (IO ByteString)

Get an IO action that reads body chunks

This is incompatible with body since body consumes all chunks.

jsonData :: (FromJSON a, MonadIO m) => ActionT m a

Parse the request body as a JSON object and return it.

If the JSON object is malformed, this sets the status to 400 Bad Request, and throws an exception.

If the JSON fails to parse, this sets the status to 422 Unprocessable Entity.

These status codes are as per https://www.restapitutorial.com/httpstatuscodes.html.

files :: Monad m => ActionT m [File]

Get list of uploaded files.

Accessing Path, Form and Query Parameters

param :: (Parsable a, MonadIO m) => Text -> ActionT m a

Deprecated: (#204) Not a good idea to treat all parameters identically. Use captureParam, formParam and queryParam instead.

Get a parameter. First looks in captures, then form data, then query parameters.

Raises an exception which can be caught by catch if parameter is not found.

If parameter is found, but parseParam fails to parse to the correct type, next is called. This means captures are somewhat typed, in that a route won't match if a correctly typed capture cannot be parsed.

params :: Monad m => ActionT m [Param]

Deprecated: (#204) Not a good idea to treat all parameters identically. Use pathParams, formParams and queryParams instead.

Get all parameters from path, form and query (in that order).

pathParam :: (Parsable a, MonadIO m) => Text -> ActionT m a

Look up a path parameter.

Raises an exception which can be caught by catch if parameter is not found. If the exception is not caught, scotty will return a HTTP error code 500 ("Internal Server Error") to the client.

If the parameter is found, but parseParam fails to parse to the correct type, next is called.

Since: 0.20

captureParam :: (Parsable a, MonadIO m) => Text -> ActionT m a

Synonym for pathParam

formParam :: (Parsable a, MonadIO m) => Text -> ActionT m a

Look up a form parameter.

Raises an exception which can be caught by catch if parameter is not found. If the exception is not caught, scotty will return a HTTP error code 400 ("Bad Request") to the client.

This function raises a code 400 also if the parameter is found, but parseParam fails to parse to the correct type.

Since: 0.20

queryParam :: (Parsable a, MonadIO m) => Text -> ActionT m a

Look up a query parameter.

Raises an exception which can be caught by catch if parameter is not found. If the exception is not caught, scotty will return a HTTP error code 400 ("Bad Request") to the client.

This function raises a code 400 also if the parameter is found, but parseParam fails to parse to the correct type.

Since: 0.20

pathParamMaybe :: (Parsable a, Monad m) => Text -> ActionT m (Maybe a)

Look up a path parameter. Returns Nothing if the parameter is not found or cannot be parsed at the right type.

NB : Doesn't throw exceptions. In particular, route pattern matching will not continue, so developers must raiseStatus or throw to signal something went wrong.

Since: 0.21

captureParamMaybe :: (Parsable a, Monad m) => Text -> ActionT m (Maybe a)

Look up a capture parameter. Returns Nothing if the parameter is not found or cannot be parsed at the right type.

NB : Doesn't throw exceptions. In particular, route pattern matching will not continue, so developers must raiseStatus or throw to signal something went wrong.

Since: 0.21

formParamMaybe :: (Parsable a, Monad m) => Text -> ActionT m (Maybe a)

Look up a form parameter. Returns Nothing if the parameter is not found or cannot be parsed at the right type.

NB : Doesn't throw exceptions, so developers must raiseStatus or throw to signal something went wrong.

Since: 0.21

queryParamMaybe :: (Parsable a, Monad m) => Text -> ActionT m (Maybe a)

Look up a query parameter. Returns Nothing if the parameter is not found or cannot be parsed at the right type.

NB : Doesn't throw exceptions, so developers must raiseStatus or throw to signal something went wrong.

Since: 0.21

pathParams :: Monad m => ActionT m [Param]

Get path parameters

captureParams :: Monad m => ActionT m [Param]

Get path parameters

formParams :: Monad m => ActionT m [Param]

Get form parameters

queryParams :: Monad m => ActionT m [Param]

Get query parameters

Modifying the Response and Redirecting

status :: MonadIO m => Status -> ActionT m ()

Set the HTTP response status.

addHeader :: MonadIO m => Text -> Text -> ActionT m ()

Add to the response headers. Header names are case-insensitive.

setHeader :: MonadIO m => Text -> Text -> ActionT m ()

Set one of the response headers. Will override any previously set value for that header. Header names are case-insensitive.

redirect :: Monad m => Text -> ActionT m a

Redirect to given URL. Like throwing an uncatchable exception. Any code after the call to redirect will not be run.

redirect "http://www.google.com"

OR

redirect "/foo/bar"

Setting Response Body

Note: only one of these should be present in any given route definition, as they completely replace the current Response body.

text :: MonadIO m => Text -> ActionT m ()

html :: MonadIO m => Text -> ActionT m ()

file :: MonadIO m => FilePath -> ActionT m ()

Send a file as the response. Doesn't set the "Content-Type" header, so you probably want to do that on your own with setHeader. Setting a status code will have no effect because Warp will overwrite that to 200 (see sendResponse).

json :: (ToJSON a, MonadIO m) => a -> ActionT m ()

Set the body of the response to the JSON encoding of the given value. Also sets "Content-Type" header to "application/json; charset=utf-8" if it has not already been set.

stream :: MonadIO m => StreamingBody -> ActionT m ()

Set the body of the response to a Source. Doesn't set the "Content-Type" header, so you probably want to do that on your own with setHeader.

raw :: MonadIO m => ByteString -> ActionT m ()

Set the body of the response to the given ByteString value. Doesn't set the "Content-Type" header, so you probably want to do that on your own with setHeader.

nested :: MonadIO m => Application -> ActionT m ()

Nest a whole WAI application inside a Scotty handler. See Web.Scotty for further documentation

Accessing the fields of the Response

getResponseHeaders :: MonadIO m => ActionT m ResponseHeaders

Access the HTTP headers of the Response

SINCE 0.21

getResponseStatus :: MonadIO m => ActionT m Status

Access the HTTP Status of the Response

SINCE 0.21

getResponseContent :: MonadIO m => ActionT m Content

Access the content of the Response

SINCE 0.21

Exceptions

raise

Arguments

:: MonadIO m

=> Text
Error text

-> ActionT m a

Deprecated: Throw an exception instead

Throw a "500 Server Error" StatusError, which can be caught with rescue.

Uncaught exceptions turn into HTTP 500 responses.

raiseStatus :: Monad m => Status -> Text -> ActionT m a

Deprecated: Use status, text, and finish instead

Throw a StatusError exception that has an associated HTTP error code and can be caught with rescue.

Uncaught exceptions turn into HTTP responses corresponding to the given status.

throw :: (MonadIO m, Exception e) => e -> ActionT m a

Throw an exception which can be caught within the scope of the current Action with catch.

If the exception is not caught locally, another option is to implement a global Handler (with defaultHandler) that defines its interpretation and a translation to HTTP error codes.

Uncaught exceptions turn into HTTP 500 responses.

rescue :: (MonadUnliftIO m, Exception e) => ActionT m a -> (e -> ActionT m a) -> ActionT m a

Deprecated: Use catch instead

Catch an exception e.g. a StatusError or a user-defined exception.

raise JustKidding `catch` (\msg -> text msg)

next :: Monad m => ActionT m a

Abort execution of this action and continue pattern matching routes. Like an exception, any code after next is not executed.

NB : Internally, this is implemented with an exception that can only be caught by the library, but not by the user.

As an example, these two routes overlap. The only way the second one will ever run is if the first one calls next.

get "/foo/:bar" $ do
  w :: Text <- pathParam "bar"
  unless (w == "special") next
  text "You made a request to /foo/special"

get "/foo/:baz" $ do
  w <- pathParam "baz"
  text $ "You made a request to: " <> w

finish :: Monad m => ActionT m a

Finish the execution of the current action. Like throwing an uncatchable exception. Any code after the call to finish will not be run.

Since: 0.10.3

defaultHandler :: Monad m => ErrorHandler m -> ScottyT m ()

Global handler for user-defined exceptions.

liftAndCatchIO :: MonadIO m => IO a -> ActionT m a

Deprecated: Use liftIO instead

Catch any synchronous IO exceptions

liftIO :: MonadIO m => IO a -> m a

Lift a computation from the IO monad. This allows us to run IO computations in any monadic stack, so long as it supports these kinds of operations (i.e. IO is the base monad for the stack).

Example

Expand

import Control.Monad.Trans.State -- from the "transformers" library

printState :: Show s => StateT s IO ()
printState = do
  state <- get
  liftIO $ print state

Had we omitted liftIO, we would have ended up with this error:

• Couldn't match type ‘IO’ with ‘StateT s IO’
 Expected type: StateT s IO ()
   Actual type: IO ()

The important part here is the mismatch between StateT s IO () and IO ().

Luckily, we know of a function that takes an IO a and returns an (m a): liftIO, enabling us to run the program and see the expected results:

> evalStateT printState "hello"
"hello"

> evalStateT printState 3
3

catch

Arguments

:: (MonadUnliftIO m, Exception e)

=> m a
action

-> (e -> m a)
handler

-> m a

Catch a synchronous (but not asynchronous) exception and recover from it.

This is parameterized on the exception type. To catch all synchronous exceptions, use catchAny.

Since: unliftio-0.1.0.0

data StatusError

Deprecated: If it is supposed to be caught, a proper exception type should be defined

E.g. when a parameter is not found in a query string (400 Bad Request) or when parsing a JSON body fails (422 Unprocessable Entity)

Constructors

StatusError Status Text

Deprecated: If it is supposed to be caught, a proper exception type should be defined

data ScottyException

Thrown e.g. when a request is too large

Constructors

RequestTooLarge

MalformedJSON ByteString Text

FailedToParseJSON ByteString Text

PathParameterNotFound Text

QueryParameterNotFound Text

FormFieldNotFound Text

FailedToParseParameter Text Text Text

Parsing Parameters

type Param = (Text, Text)

class Parsable a where

Minimum implemention: parseParam

Minimal complete definition

parseParam

Methods

parseParam :: Text -> Either Text a

Take a Text value and parse it as a, or fail with a message.

parseParamList :: Text -> Either Text [a]

Default implementation parses comma-delimited lists.

parseParamList t = mapM parseParam (T.split (== ',') t)

readEither :: Read a => Text -> Either Text a

Useful for creating Parsable instances for things that already implement Read. Ex:

instance Parsable Int where parseParam = readEither

Types

data RoutePattern

type File = (Text, FileInfo ByteString)

data Content

Constructors

ContentBuilder Builder

ContentFile FilePath

ContentStream StreamingBody

ContentResponse Response

type Kilobytes = Int

type ErrorHandler m = Handler (ActionT m) ()

Specializes a Handler to the ActionT monad

data Handler (m :: Type -> Type) a

Generalized version of Handler

Constructors

Exception e => Handler (e -> m a)

Monad Transformers

data ScottyT m a

data ActionT m a

data ScottyState m

defaultScottyState :: ScottyState m