Web.Scotty.Internal.Types

Documentation

data Options

Constructors

Options

Fields

verbose :: Int

0 = silent, 1(def) = startup banner

settings :: Settings

Warp Settings Note: to work around an issue in warp, the default FD cache duration is set to 0 so changes to static files are always picked up. This likely has performance implications, so you may want to modify this for production servers using setFdCacheDuration.

type Middleware m = Application m -> Application m

type Application m = Request -> m Response

data ScottyState e m

Constructors

ScottyState

Fields

middlewares :: [Middleware]

routes :: [Middleware m]

handler :: ErrorHandler e m

addMiddleware :: Middleware -> ScottyState e m -> ScottyState e m

addRoute :: Middleware m -> ScottyState e m -> ScottyState e m

addHandler :: ErrorHandler e m -> ScottyState e m -> ScottyState e m

newtype ScottyT e m a

Constructors

ScottyT

Fields

runS :: State (ScottyState e m) a

data ActionError e

Constructors

Redirect Text

Next

Finish

ActionError e

class ScottyError e where

In order to use a custom exception type (aside from Text), you must define an instance of ScottyError for that type.

Methods

stringError :: String -> e

showError :: e -> Text

type ErrorHandler e m = Maybe (e -> ActionT e m ())

type Param = (Text, Text)

type File = (Text, FileInfo ByteString)

data ActionEnv

Constructors

Env

Fields

getReq :: Request

getParams :: [Param]

getBody :: IO ByteString

getBodyChunk :: IO ByteString

getFiles :: [File]

data RequestBodyState

Constructors

BodyUntouched

BodyCached ByteString [ByteString]

BodyCorrupted

data BodyPartiallyStreamed

Constructors

BodyPartiallyStreamed

data Content

Constructors

ContentBuilder Builder

ContentFile FilePath

ContentStream StreamingBody

data ScottyResponse

Constructors

SR

Fields

srStatus :: Status

srHeaders :: ResponseHeaders

srContent :: Content

newtype ActionT e m a

Constructors

ActionT

Fields

runAM :: ExceptT (ActionError e) (ReaderT ActionEnv (StateT ScottyResponse m)) a

data RoutePattern

Constructors

Capture Text

Literal Text

Function (Request -> Maybe [Param])